backend/requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.12
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2

